Split test suite
